+++ openhands/server/config/server_config.py
@@ -16,7 +16,7 @@ class ServerConfig(ServerConfigInterface):
         'openhands.storage.settings.file_settings_store.FileSettingsStore'
     )
     secret_store_class: str = (
-        'openhands.storage.settings.file_secrets_store.FileSecretsStore'
+        'openhands.storage.secrets.file_secrets_store.FileSecretsStore'
     )
     conversation_store_class: str = (
         'openhands.storage.conversation.file_conversation_store.FileConversationStore'
